+++ helpers/dev.js
@@ -447,11 +447,17 @@ export function timer(startTime: Date): string {
 }
 
 /**
+ * A special logger that logs the time it takes to execute a function, or a certain stage of a function, that this is called from.
+ * Assumes that `const startTime = new Date()` is included earlier in the function.
+ * If separate _logTimer setting is true, then it will log the time, irrespective of the main _logLevel setting.
+ * But if warningTrigger (in milliseconds)is exceeded, then this will log with a warning, irrespective of _logTimer or _logLevel settings.
  * @author @jgclark
- * @param {*} startTime - the date object from when timer started (using Date.now())
- * @returns {string}
+ * @param {string} functionName - to display after time in log line
+ * @param {Date} startTime - the date object from when timer started (using Date.now())
+ * @param {string} explanation - (optional) text to display after the duration in log line
+ * @param {number} warningTrigger - optional duration in milliseconds: if the timer is more than this it will log with added warning symbol.
  */
-export function logTimer(functionName: string, startTime: Date, explanation: string = '', warningTrigger?: number): string {
+export function logTimer(functionName: string, startTime: Date, explanation: string = '', warningTrigger?: number): void {
   // $FlowIgnore[unsafe-arithmetic]
   const difference = new Date() - startTime
   const diffTimeText = `${difference.toLocaleString()}ms`
@@ -467,7 +473,6 @@ export function logTimer(functionName: string, startTime: Date, explanation: str
       console.log(msg)
     }
   }
-  return ''
 }
 
 /**