+++ jgclark.Dashboard/src/react/components/Header/Header.jsx
@@ -11,7 +11,7 @@
 import React from 'react'
 // import { getFeatureFlags } from '../../shared.js'
 import { createDashboardSettingsItems } from '../../../dashboardSettings.js'
-import type { TNotePlanConfig } from '../../../types.js'
+// import type { TNotePlanConfig } from '../../../types.js'
 import { useSettingsDialogHandler } from '../../customHooks/useSettingsDialogHandler.jsx'
 import DropdownMenu from '../DropdownMenu.jsx'
 import SettingsDialog from '../SettingsDialog.jsx'
@@ -75,7 +75,7 @@ const Header = ({ lastFullRefresh }: Props): React$Node => {
   //----------------------------------------------------------------------
   // Constants
   //----------------------------------------------------------------------
-  const { /*dashboardSettings: pluginDataSettings, */ notePlanSettings, logSettings } = pluginData
+  const { /*dashboardSettings: pluginDataSettings, notePlanSettings, */ logSettings } = pluginData
 
   // const { FFlag_DashboardSettings } = getFeatureFlags(notePlanSettings, dashboardSettings)